% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgemat.R
\name{edgemat}
\alias{edgemat}
\title{Generate the edge matrix using R}
\usage{
edgemat(p, pr)
}
\arguments{
\item{p}{Variable dimension}

\item{pr}{Probability matrix}
}
\value{
a matrix consisting of edge information
}
\description{
Generate relationships between variables with a set probability
}
\examples{
\dontrun{
p <- 100
x <- runif(p,0,1)
y <- runif(p,0,1)
node <- cbind(x,y)
distance <- as.matrix(dist(node))
pr <- matrix(0,nrow = p,ncol = p)
for (i in 1:(p-1)) {
for (j in (i+1):p) {
d <- distance[i,j]
pr[i,j] <- dnorm(d * sqrt(p))
}}
e <- edgemat(p,pr)}
}
