% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gram.R
\name{Gram}
\alias{Gram}
\title{Generate the Gram matrix using R}
\usage{
Gram(m)
}
\arguments{
\item{m}{Input matrix}
}
\value{
a matrix consisting of inner products
}
\description{
Generate the Gram matrix which consists of the inner products of vectors
}
\examples{
\dontrun{
data(data)
g <- Gram(state.x77)
}
}
